requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
structlog==23.1.0
slowapi==0.1.8
pyjwt[crypto]==2.8.0
redis==5.0.0
anthropic==0.3.11
langchain==0.0.300
zep-python==1.1.0
tiktoken==0.5.1
requests==2.31.0
pandas==2.1.0
yfinance==0.2.28
numpy==1.25.2
aiohttp>=3.13.3
python-multipart==0.0.6
python-dotenv==1.0.0
asyncio==3.4.3
aioredis==2.0.1
sqlalchemy==2.0.20
asyncpg==0.28.0
alembic==1.12.0
psutil==5.9.5
pytz==2023.3
scipy==1.11.2
statsmodels==0.14.0
scikit-learn==1.3.0
torch==2.0.1
transformers>=4.50.0
websockets==11.0.3
alpaca-trade-api==3.1.1
opencv-python-headless==4.8.0.76
pillow==10.0.0
rasterio==1.3.8
folium==0.14.0
geopandas==0.13.2
shapely==2.0.1
pyproj==3.6.0
pyshp==2.3.1
fiona==1.9.4.post1
h3==3.7.6
apscheduler==3.10.1
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
vaderSentiment==3.3.2
blazingmq-sdk-python==1.0.1
sgp4==2.23
feedparser==6.0.10
httpx==0.24.1
aiofiles==23.2.1
# Bloomberg Performance Tools (Needs C++ compiler for native setup)
# memray
# pystack
